Inline module into __init__.py
