--- {{ cookiecutter.directory_name }}/{{ cookiecutter.package_name }}/module.py
@@ -1,32 +1,0 @@
-from typing import Any, Dict
-
-import attr
-from synapse.module_api import ModuleApi
-
-
-@attr.s(auto_attribs=True, frozen=True)
-class {{ cookiecutter.module_class_name }}Config:
-    pass
-
-
-class {{ cookiecutter.module_class_name }}:
-    def __init__(self, config: {{ cookiecutter.module_class_name }}Config, api: ModuleApi):
-        # Keep a reference to the config and Module API
-        self._api = api
-        self._config = config
-
-    @staticmethod
-    def parse_config(config: Dict[str, Any]) -> {{ cookiecutter.module_class_name }}Config:
-        # Parse the module's configuration here.
-        # If there is an issue with the configuration, raise a
-        # synapse.module_api.errors.ConfigError.
-        #
-        # Example:
-        #
-        #     some_option = config.get("some_option")
-        #     if some_option is None:
-        #          raise ConfigError("Missing option 'some_option'")
-        #      if not isinstance(some_option, str):
-        #          raise ConfigError("Config option 'some_option' must be a string")
-        #
-        return {{ cookiecutter.module_class_name }}Config()
